Remove EDH pages api methods & tests
